requirements.txt
flask>=3.0
requests>=2.31
python-dotenv>=1.0
beautifulsoup4>=4.12
psycopg2-binary>=2.9
